requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
decorator==4.4.2
future==0.18.2
idna==2.9
numpy==1.18.1
pandas==1.0.3
pycparser==2.20
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
python-dateutil==2.8.1
python-telegram-bot==12.6.1
pytz==2020.1
requests==2.7.0
six==1.14.0
telebot==0.0.3
tornado==6.0.4
urllib3==1.25.9
